Required files/Completed assignment
Created/added required files to run this django app.
Set up templates/static folders correctly to have them in use with the app.
Practiced linking css to html files.
Used render() for the first time on time_display.views.py to render a template with a dict and displayed the value of a key on that template.